+++ src/lec_moving_framework.py
@@ -6,7 +6,7 @@
 #    By: daniloceano <danilo.oceano@gmail.com>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/12/19 17:32:55 by daniloceano       #+#    #+#              #
-#    Updated: 2023/12/21 11:02:51 by daniloceano      ###   ########.fr        #
+#    Updated: 2023/12/21 16:10:46 by daniloceano      ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,7 +27,7 @@ from EnergyContents import EnergyContents
 from ConversionTerms import ConversionTerms
 from BoundaryTerms import BoundaryTerms
 from GenerationDissipationTerms import GenerationDissipationTerms
-from BoxData import BoxData
+from box_data import BoxData
 from BudgetResidual import calc_budget_diff, calc_residuals
 
 def create_terms_dict(args):
